+++ .gitignore
@@ -3,4 +3,5 @@ out
 tmp
 storage/
 *secret*
+*tmux*
 *\.ext4
