+++ Cargo.toml
@@ -24,21 +24,10 @@ path = "bins/server/main.rs"
 name = "fc_wrapper"
 path = "bins/fc_wrapper/main.rs"
 
-[[bin]]
-name = "webhook"
-path = "bins/webhook/main.rs"
-
 [lib]
 
 [dependencies]
-flate2 = "^1.0.22"
-tar = "^0.4.37"
 ctrlc = "^3.2.0"
-ring = "^0.16.20"
-curl = "^0.4.41"
-github-types = "=0.1.1"
-http = "~0.2"
-httparse = "^1.3.4"
 reqwest = { version = "0.11", features = ["blocking"] }
 labeled = { git = "git://github.com/alevy/labeled" }
 bytes = "1.1.0"
@@ -51,10 +40,8 @@ env_logger = "^0.9.0"
 serde = {version = "1.0.102", features = ["derive"]}
 serde_yaml = "0.8.11"
 serde_json = "*"
-shellexpand = "1.0.0"
-nix = "0.16.0"
-cgroups = "0.1.0"
 vmm = { path = "firecracker/vmm", features = ["vsock"]}
+cgroups = "0.1.0"
 memory_model = { path = "firecracker/memory_model" }
 fc_util = { path = "firecracker/fc_util" }
 sys_util = { path = "firecracker/sys_util" }
