+++ src/lib.rs
@@ -11,8 +11,6 @@ pub mod vm;
 pub mod syscalls;
 pub mod metrics;
 pub mod firecracker_wrapper;
-pub mod server;
-pub mod app;
 
 use std::string::String;
 use std::fs::{self, File};
