+++ src/request.rs
@@ -1,6 +1,12 @@
 use std::io::{Error, ErrorKind, Write, Read};
 use serde::{Deserialize, Serialize};
 use serde_json::Value;
+                
+#[derive(Debug, Serialize, Deserialize)]
+pub struct Response {
+    pub user_id: u64,
+    pub payload: Value,
+}
 
 #[derive(Debug, Serialize, Deserialize)]
 pub struct Request {
