+++ webhook/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "webhook"
+version = "0.1.0"
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+ring = "^0.16.20"
+bytes = "^1.1.0"
+http = "~0.2"
+httparse = "^1.3.4"
+serde = {version = "1.0.102", features = ["derive"]}
+serde_yaml = "*"
+serde_json = "*"
+log = { version = "0.4", features = ["max_level_debug", "release_max_level_warn"] }
+env_logger = "^0.9.0"
+clap = "^2.33.0"
+github-types = "^0.1.1"
+snapfaas = { path = "../." }
