+++ webhook/src/main.rs
@@ -1,17 +1,18 @@
 use clap::{App, Arg};
-use snapfaas::server;
 
 mod app;
 mod config;
+mod server;
 
 fn main() -> Result<(), std::io::Error> {
     env_logger::init();
 
     let matches = App::new("webhook server")
         .arg(
-            Arg::with_name("snapfaas config")
-                .short("c")
-                .long("snapfaas_config")
+            Arg::with_name("snapfaas address")
+                .short("p")
+                .long("snapfaas_address")
+                .value_name("[ADDR:]PORT")
                 .takes_value(true)
                 .required(true)
                 .help("Path to snapfaas config YAML file"),
@@ -20,6 +21,7 @@ fn main() -> Result<(), std::io::Error> {
             Arg::with_name("app config")
                 .short("a")
                 .long("app_config")
+                .value_name("YAML")
                 .takes_value(true)
                 .required(true)
                 .help("Path to app config YAML file"),
@@ -33,22 +35,13 @@ fn main() -> Result<(), std::io::Error> {
                 .required(true)
                 .help("Address to listen on"),
         )
-        .arg(Arg::with_name("total memory")
-                .long("mem")
-                .takes_value(true)
-                .value_name("MB")
-                .required(true)
-                .help("Total memory available for all VMs")
-        )
         .get_matches();
 
     let app = app::App::new(matches.value_of("app config").unwrap());
     let server = server::Server::new(
-        matches.value_of("total memory").unwrap().parse::<usize>().expect("Total memory is not a valid integer"),
-        matches.value_of("snapfaas config").unwrap(),
+        matches.value_of("snapfaas address").unwrap().to_string(),
         matches.value_of("listen").unwrap(),
         app
     );
-    server.set_ctrlc_handler();
     server.run()
 }
